Example showing UInt cake-like refactor
This refactors UInt into two private traits: one for macros and one for
method implementations. This is one way to cleanup Bits while moving the
body of sealed classes (Bits, UInt, SInt, FixedPoint, Analog) into
separate files.
